tests almost done

+++ webstore/src/main/java/pl/spring/demo/controller/BookController.java
@@ -5,20 +5,14 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.InitBinder;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.servlet.ModelAndView;
 
 import pl.spring.demo.constants.ModelConstants;
 import pl.spring.demo.constants.ViewNames;
-import pl.spring.demo.repository.BookRepository;
 import pl.spring.demo.service.BookService;
 import pl.spring.demo.to.BookTo;
 
@@ -35,7 +29,6 @@ public class BookController {
 	@Autowired
 	private BookService bookService;
 
-
 	@RequestMapping
 	public String list(Model model) {
 		List<BookTo> books = bookService.findAllBooks();
@@ -43,24 +36,21 @@ public class BookController {
 		return ViewNames.BOOKS;
 	}
 
-
-
-
-	
 	@RequestMapping(value = "/book", method = RequestMethod.GET)
 	public String findBookDetails(@RequestParam Long id, Model model) {
 		BookTo book = bookService.findBooksById(id);
-		model.addAttribute(ModelConstants.BOOK,book);
+		model.addAttribute(ModelConstants.BOOK, book);
 		return ViewNames.BOOK;
 	}
-	
+
 	@RequestMapping(value = "/delete")
 	public String deleteBook(@RequestParam Long id, Model model) {
-		bookService.deleteBook(id);
+		if (bookService.findBooksById(id) == null)
+			return ViewNames.NOTFOUND;
+		else
+			bookService.deleteBook(id);
 		return ViewNames.DELETE;
 	}
-	
-	
 
 	/**
 	 * Binder initialization
